Host consent page on a Cloudflare Worker (free); extract shared template
Supabase EFs can't serve text/html on the default domain, so the consent
page moves to a free Cloudflare Worker (workers.dev, no owned domain
needed). Markup extracted to _shared/consent-page/renderConsentPage() as
the single source; the Cloudflare Worker and the (retained, custom-domain-
only) cerefox-oauth-consent EF both render it. bundle_server_assets ships
consent-page so the EF deploy still resolves.

Design updated with the consent-host decision and OSS-framing open items
(optional feature scoping; the "is the public anon key really a secret?"
question; ChatGPT-via-MCP caveat).
