+++ cloudflare/cerefox-consent/wrangler.toml
@@ -1,0 +1,14 @@
+# Cloudflare Worker: Cerefox OAuth consent page (design §4.2-B).
+# Deploy:  cd cloudflare/cerefox-consent && npx wrangler deploy
+# First run prompts `wrangler login` (browser). Free Workers plan; gives a
+# https://cerefox-consent.<your-subdomain>.workers.dev URL.
+name = "cerefox-consent"
+main = "src/index.ts"
+compatibility_date = "2024-11-01"
+workers_dev = true
+
+# Both values are PUBLIC (project URL + anon JWT). Fill in your project's, or
+# override per-deploy with:  npx wrangler deploy --var SUPABASE_URL:... --var SUPABASE_ANON_KEY:...
+[vars]
+SUPABASE_URL = "https://YOUR_PROJECT_REF.supabase.co"
+SUPABASE_ANON_KEY = "REPLACE_WITH_YOUR_LEGACY_ANON_JWT"
